GCP Deployment/requirements.txt
Flask>=2.0.1
gensim>=4.3.2
numpy>=1.21.0
gunicorn>=20.1.0
google-cloud-storage>=1.42.0
